Create config package for openscap-plugin
Although it is not yet defined all options to be used in a plugin
configuration file, it is already clear that a configuration file is
necessary. This commit includes the config package with code to read
and validate configuration while also sanitize values. Unit tests
specific for the configuration file are not implemented now since the
configuration file structure is still under development and
considerable changes are expected.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>