feat: auto-review dependabot pull requests
Identify the version change and classify the risk for based on major,
minor or patch updates. If the risk is low or medium, auto-approve the
PR and wait for a maintainer (human) to also review. If the risk is
high, two maintainers need to approve. If the PR indicates a security
update, add a label for faster triage.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>