feat: adds configuration for defining assessment scope in the Assessment Plan (#134)
* feat: introducing the write function to update plan

* feat: populating the config file in .json format

* fix: removing the populated config files

* feat: includes a function to populate config.yaml

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

* feat: revising structure of config.yml

* fix: updating to only populate yaml in StdOut
This commit will print the testdata to StdOut
as key: value pairs. The content can be updated
in the config.yaml file in the user workspace.

* feat: update command

Signed-off-by: d10n <d10n@redhat.com>

* feat: print default assessment plan filter

Signed-off-by: d10n <d10n@redhat.com>

* chore: reduce indentation

Signed-off-by: d10n <d10n@redhat.com>

* feat: assessment plan control filter

Signed-off-by: d10n <d10n@redhat.com>

* refactor: adds AssessmentScope type around plan filtering

The goal of the AssessmentScope type is to add domain-level clarity
around the intention for assessment plan filering. This is moved to a new
package to allow for expansion without significant rewrites and to keep the
CLI layer focused on I/O.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: updating planDryRun function

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

This commit edits the planDryRun command reference
to the ControlImplementations. The applicable
controls should be specific to FrameworkID. The
FrameworkID property of the AssessmentScope is
referenced using the Source field on the
ControlImplementationSet.

* feat: updating location of filtered plan

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

* fix: updating the framework control inclusion

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

This commit compares the name of the frameworkID with the
Source on the ControlImplementationSet. Then, the
loop checks for the control ids that are part
of the set of controls that apply to the FrameworkID
and should be included in the assessment plan.

* feat: adding controls by FrameworkID

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

This commit compares the name of the frameworkID with the
Source on the ControlImplementationSet. Then, the
loop checks for the control ids that are part
of the set of controls that apply to the FrameworkID
and should be included in the assessment plan.

# Conflicts:
#	cmd/complytime/cli/plan.go

* fix: apply fixes to broken build

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat: remove activities for out-of-scope controls

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

This commit removes the RelatedControls fields
for the OSCAL Assessment Plan activities. The
activites with out-of-scope controls will be marked
"skipped."

* feat: introducing skipped property

fix: updating unit test for scope

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

* feat: unit tests for filtering controls

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

* feat: adding controls by FrameworkID

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

This commit compares the name of the frameworkID with the
Source on the ControlImplementationSet. Then, the
loop checks for the control ids that are part
of the set of controls that apply to the FrameworkID
and should be included in the assessment plan.

# Conflicts:
#	cmd/complytime/cli/plan.go

* fix: apply fixes to broken build

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates plan examples description and formatting

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updating config reference for plan command

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>

fix: apply fixes to broken build

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: add test to reproduce duplicate error

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: updates AssessmentScope created to deduplicated controlIds

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* refactor: add helper type to improve consistency

includeControls sets with map[string]bool were implemented
multiple times. This adds an includeControlsSet helper type
for readabilty and consistency throughout scope.go

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore(deps): bump c2p version to latest

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

---------

Signed-off-by: Hannah Braswell <hbraswel@redhat.com>
Signed-off-by: d10n <d10n@redhat.com>
Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>
Co-authored-by: Hannah Braswell <hbraswel@redhat.com>
Co-authored-by: d10n <d10n@redhat.com>